requirements.txt
# requirements.txt

# Core FastAPI and Server Libraries
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6

# AI and LLM Libraries
openai==1.12.0
tiktoken==0.5.2
tenacity==8.2.3

# Database Libraries
sqlalchemy==2.0.23

# Advanced Features & Monitoring
redis==5.0.1
psutil==5.9.6
numpy==1.24.3
scikit-learn==1.3.2
aiofiles==23.2.1
aiohttp==3.9.1

# CORS support
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4